apply changes following MFitz PR comments
